Smart Filter - Supply Exact multiplier feature
Supply Exact mode with a Smart Filter now displays the quantity bar.

Instead of a number of items, this is used as a multiplier so you can
send exact multiples of ingredients (as long as the Robot Arm can move
that many items at once). Useful for stuff like Processing Arrays.

For example, a Smart Filter Supplying Exact in Electrolyzer Mode with
the quantity 16 will send 16 graphite dusts per operation (16*1 dust)
or exactly 624 bauxite dust (16*39 dusts).

If the amount exceeds the maximum transfer rate of the Robot Arm, or
there is not enough material, or not enough room to move all the items,
then the Robot Arm will not do anything.