Add "Jammed" state as a contingency for reformed structures
Since recipes no longer clear when a structure is broken (so long as the
controller isn't broken, anyway), a contingency is needed to avoid
undefined behavior when the player reforms the structure with less
output space than was available when the recipe was initialized.

Structures will become "Jammed" if at the time of recipe completion,
there is insufficient space for all outputs as computed during recipe
setup. This behavior will enforce additional constraints via
RecipeMapMultiblockController::checkRecipe(Recipe, boolean).

While Jammed, crafting will pause at 100% and the controller GUI will
indicate the structure cannot complete the recipe as configured.

When the player reforms the structure appropriately, the
Jammed status will clear and the craft will complete.