Fix Superconductors not getting cable recipes
Also refactored/documented some related code.

This bug happened because Superconductor is not an IngotMaterial so it
was skipped over by the class-based processing handler.

Created a handler registry method for Predicate based matching for a
given OrePrefix and Material target that carries a payload to apply to
matches. This is used to match Superconductor and provide its
WireProperties as it's not directly contained in its Material type.