Corrections for GTUtilsTest
The previous test data did not match the actual behavior of the existing
implementation of getTierByVoltage, resulting in that test failing.

The existing behavior of that method is to find the highest tier whose
voltage is less than or equal to the target, not greater than or equal
to the target like GTCEu (whose algorithm I am testing to potentially
replace the existing implementation).

- Modified voltageTestArgs to reflect the existing behavior so that the
  test_voltage_tier test passes.

- Modified the experimental logic slightly so that its behavior is
  identical to the current behavior.